@@ -106,7 +106,7 @@ public class ProjectJavaSourcesImpl implements ProjectJavaSources {
     @Override
     public List<MethodCall> findMethodCalls(String methodPattern) {
         List<MethodCall> matches = new ArrayList<>();
-        FindMethods findMethods = new FindMethods(methodPattern,true, null);
+        FindMethods findMethods = new FindMethods(methodPattern,true);
         MethodMatcher methodMatcher = new MethodMatcher(methodPattern);
         find(findMethods).stream()
                 .map(m -> list().stream().filter(js -> js.getResource().getId().equals(m.getId())).findFirst().get())
@@ -132,7 +132,7 @@ public class ProjectJavaSourcesImpl implements ProjectJavaSources {
 
     @Override
     public List<? extends JavaSource> findClassesUsingType(String type) {
-        UsesType<ExecutionContext> usesType = new UsesType<>(type);
+        UsesType<ExecutionContext> usesType = new UsesType<>(type, true);
         GenericOpenRewriteRecipe<UsesType<ExecutionContext>> recipe = new GenericOpenRewriteRecipe<>(() -> usesType);
         return find(recipe).stream()
                 .filter(RewriteSourceFileHolder.class::isInstance)
