@@ -18,6 +18,7 @@ package org.springframework.sbm.java.impl;
 import lombok.Getter;
 import org.jetbrains.annotations.NotNull;
 import org.openrewrite.ExecutionContext;
+import org.openrewrite.SourceFile;
 import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.java.JavaParser;
 import org.openrewrite.java.marker.JavaSourceSet;
@@ -26,10 +27,12 @@ import org.springframework.sbm.project.resource.SbmApplicationProperties;
 import org.springframework.sbm.scopes.annotations.ScanScope;
 import org.springframework.stereotype.Component;
 
+import java.net.URI;
 import java.nio.file.Path;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Stream;
 
 @Component
 @ScanScope
@@ -58,8 +61,15 @@ public class RewriteJavaParser implements JavaParser {
         return builder.build();
     }
 
-    @Override
-    public List<J.CompilationUnit> parseInputs(Iterable<Input> sources, @Nullable Path relativeTo, ExecutionContext ctx) {
+    public List<J.CompilationUnit> parseInputsToCompilationUnits(Iterable<Input> sources, @Nullable Path relativeTo, ExecutionContext ctx) {
+        reset();
+        return parseInputs(sources, relativeTo, executionContext)
+                .filter(J.CompilationUnit.class::isInstance)
+                .map(J.CompilationUnit.class::cast)
+                .toList();
+    }
+
+    public Stream<SourceFile> parseInputs(Iterable<Input> sources, @Nullable Path relativeTo, ExecutionContext ctx) {
         reset();
         return this.javaParser.parseInputs(sources, relativeTo, ctx);
     }
@@ -70,27 +80,22 @@ public class RewriteJavaParser implements JavaParser {
     }
 
     @Override
+    public JavaParser reset(Collection<URI> uris) {
+        return javaParser.reset();
+    }
+
+    @Override
     public void setClasspath(Collection<Path> classpath) {
         this.javaParser.setClasspath(classpath);
     }
 
-    @Override
-    public void setSourceSet(String sourceSet) {
-        this.javaParser.setSourceSet(sourceSet);
-    }
-
-    @Override
-    public JavaSourceSet getSourceSet(ExecutionContext ctx) {
-        return this.javaParser.getSourceSet(ctx);
-    }
-
-    public List<J.CompilationUnit> parse(List<Path> javaResources, ExecutionContext executionContext) {
+    public Stream<SourceFile> parse(List<Path> javaResources, ExecutionContext executionContext) {
         reset();
         return this.parse(javaResources, null, executionContext);
     }
 
     @Override
-    public List<J.CompilationUnit> parse(String... sources) {
+    public Stream<SourceFile> parse(String... sources) {
         return this.parse(executionContext, sources);
     }
 }
