@@ -42,7 +42,12 @@ public class ChangeMethodReturnTypeRecipe extends Recipe {
 	}
 
 	@Override
-	protected TreeVisitor<?, ExecutionContext> getVisitor() {
+	public String getDescription() {
+		return getDisplayName();
+	}
+
+	@Override
+	public TreeVisitor<?, ExecutionContext> getVisitor() {
 		return new JavaIsoVisitor<>() {
 
 			@Override
