@@ -156,7 +156,12 @@ public class VisitorUtils {
         }
 
         @Override
-        protected TreeVisitor<?, ExecutionContext> getVisitor() {
+        public String getDescription() {
+            return getDisplayName();
+        }
+
+        @Override
+        public TreeVisitor<?, ExecutionContext> getVisitor() {
             return new JavaIsoVisitor<ExecutionContext>() {
 
                 @Override
@@ -169,7 +174,8 @@ public class VisitorUtils {
                             removeMarker(expression, marker);
                             MethodDeclaration method = getCursor().firstEnclosing(MethodDeclaration.class);
                             if (method != null) {
-                                doAfterVisit(new ChangeMethodReturnTypeRecipe(m -> m.getId().equals(method.getId()), marker.getExpression(), marker.getImports()));
+                                ChangeMethodReturnTypeRecipe changeMethodReturnTypeRecipe = new ChangeMethodReturnTypeRecipe(m -> m.getId().equals(method.getId()), marker.getExpression(), marker.getImports());
+                                doAfterVisit(changeMethodReturnTypeRecipe.getVisitor());
                             }
                         }
                     }
