@@ -17,10 +17,10 @@ package org.springframework.sbm.project.parser;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.Resource;
-import org.springframework.sbm.utils.LinuxWindowsPathUnifier;
 import org.springframework.sbm.common.util.OsAgnosticPathMatcher;
 import org.springframework.sbm.project.resource.SbmApplicationProperties;
 import org.springframework.sbm.project.resource.ResourceHelper;
+import org.springframework.sbm.utils.LinuxWindowsPathUnifier;
 import org.springframework.stereotype.Service;
 import org.springframework.util.PathMatcher;
 
