@@ -15,18 +15,24 @@
  */
 package org.springframework.sbm.project.resource;
 
+import lombok.RequiredArgsConstructor;
+import org.openrewrite.ExecutionContext;
+import org.springframework.sbm.engine.recipe.RewriteMigrationResultMerger;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 
 @Component
+@RequiredArgsConstructor
 public class ProjectResourceSetHolder {
     private ProjectResourceSet projectResourceSet;
+    private final ExecutionContext executionContext;
+
     public void setProjectResourceSet(ProjectResourceSet projectResourceSet) {
         this.projectResourceSet = projectResourceSet;
     }
 
     public ProjectResourceSet getProjectResourceSet() {
-        return projectResourceSet == null ? new ProjectResourceSet(new ArrayList<>()) : projectResourceSet;
+        return projectResourceSet == null ? new ProjectResourceSet(new ArrayList<>(), executionContext) : projectResourceSet;
     }
 }
