@@ -39,7 +39,7 @@ public class StringProjectResource extends RewriteSourceFileHolder<PlainText> {
      */
     public StringProjectResource(Path absolutePath, ExecutionContext executionContext) {
         //super(new RewriteSourceFileHolder<>(new PlainTextParser().parse(List.of(absolutePath), null, new RewriteExecutionContext()).get(0)));
-        super(absolutePath, new PlainTextParser().parse(List.of(absolutePath), null, executionContext).get(0));
+        super(absolutePath, (PlainText) new PlainTextParser().parse(List.of(absolutePath), null, executionContext).toList().get(0));
         this.executionContext = executionContext;
     }
 
@@ -48,7 +48,7 @@ public class StringProjectResource extends RewriteSourceFileHolder<PlainText> {
      */
     public StringProjectResource(Path projectRoot, Path absolutePath, String content, ExecutionContext executionContext) {
         // FIXME: absolutePath, sourcePath, modulePath ?!
-        super(projectRoot, new PlainTextParser().parse(content).get(0).withSourcePath(projectRoot.relativize(absolutePath)));
+        super(projectRoot, new PlainTextParser().parse(content).toList().get(0).withSourcePath(projectRoot.relativize(absolutePath)));
         this.content = content;
         this.executionContext = executionContext;
         markAsChanged();
